refactor(params): extract defaults into a separate file
